// types/property.tsx
export type Property = {
        id: number;
        image: string;
        firstName: string;
        lastName: string;
        location: string;
        experience: number;
        wage: number;
        wageFrequency: string;
        currency: string;
        favorite: boolean;
        tag: string[];
}